M+awsb/requirements.txt
asgiref==3.4.1
bokeh==2.4.3
certifi==2021.10.8
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==3.2.13
django-crispy-forms==1.13.0
django-filter==21.1
django-related-admin==1.0.1
django-slick-reporting==0.6.3
djangorestframework==3.13.1
flatdict==4.0.1
greenlet==1.1.2
idna==3.3
Jinja2==3.0.3
MarkupSafe==2.0.1
mysqlclient==2.1.0
numpy==1.21.5
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
pyparsing==3.0.6
python-dateutil==2.8.2
python-decouple==3.6
pytz==2021.3
PyYAML==6.0
requests==2.27.1
simplejson==3.17.6
six==1.16.0
SQLAlchemy==1.4.37
sqlparse==0.4.2
tornado==6.1
typing-extensions==4.0.1
urllib3==1.26.9
